Extract enum generation into its own class
